Move getPatronName down into the MARC namespace
